Add users details page (only admins can see)

+++ db/schema.rb
@@ -14,8 +14,8 @@ ActiveRecord::Schema[7.0].define(version: 20_221_012_060_146) do
   create_table 'active_storage_attachments', force: :cascade do |t|
     t.string 'name', null: false
     t.string 'record_type', null: false
-    t.bigint 'record_id', null: false
-    t.bigint 'blob_id', null: false
+    t.integer 'record_id', null: false
+    t.integer 'blob_id', null: false
     t.datetime 'created_at', null: false
     t.index ['blob_id'], name: 'index_active_storage_attachments_on_blob_id'
     t.index %w[record_type record_id name blob_id], name: 'index_active_storage_attachments_uniqueness',
@@ -35,7 +35,7 @@ ActiveRecord::Schema[7.0].define(version: 20_221_012_060_146) do
   end
 
   create_table 'active_storage_variant_records', force: :cascade do |t|
-    t.bigint 'blob_id', null: false
+    t.integer 'blob_id', null: false
     t.string 'variation_digest', null: false
     t.index %w[blob_id variation_digest], name: 'index_active_storage_variant_records_uniqueness', unique: true
   end
